# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  requests.CheckInRequest:
    properties:
      device_info:
        maxLength: 255
        type: string
      location:
        maxLength: 255
        type: string
      qr_payload:
        minLength: 10
        type: string
    required:
    - qr_payload
    type: object
...
  requests.ManualCheckInRequest:
    properties:
      device_info:
        maxLength: 255
        type: string
      location:
        maxLength: 255
        type: string
      ticket_code:
        maxLength: 32
        minLength: 6
        type: string
    required:
    - ticket_code
    type: object
...
  responses.BrandDashboardResponse:
    properties:
      active_partnerships:
        type: integer
      available_events:
        type: integer
      pending_requests:
        type: integer
    type: object
...
  responses.CheckInRecordResponse:
    properties:
      event_id:
        type: string
      id:
        type: string
      location:
        type: string
      method:
        type: string
      scanned_at:
        type: string
      scanned_by:
        type: string
      ticket_id:
        type: string
    type: object
  responses.CheckInResponse:
    properties:
      already_scanned:
        type: boolean
      checked_in_at:
        type: string
      scanned_by:
        type: string
      ticket:
        $ref: '#/definitions/responses.TicketResponse'
    type: object
  responses.CheckInStatsResponse:
    properties:
      checked_in:
        type: integer
      checked_in_percent:
        type: number
      event_id:
        type: string
      remaining:
        type: integer
      total_tickets:
        type: integer
    type: object
...
  responses.ErrorDetail:
    properties:
      code:
        type: string
      details: {}
      message:
        type: string
    type: object
  responses.ErrorResponse:
    properties:
      error:
        $ref: '#/definitions/responses.ErrorDetail'
      request_id:
        type: string
      success:
        type: boolean
      timestamp:
        type: string
    type: object
...
  responses.EventSummaryResponse:
    properties:
      banner_url:
        type: string
      category:
        type: string
      created_at:
        type: string
      currency:
        type: string
      end_at:
        type: string
      event_type:
        type: string
      id:
        type: string
      is_featured:
        type: boolean
      organizer_id:
        type: string
      slug:
        type: string
      start_at:
        type: string
      status:
        type: string
      summary:
        type: string
      thumbnail_url:
        type: string
      timezone:
        type: string
      title:
        type: string
      venue_city:
        type: string
      venue_country:
        type: string
      visibility:
        type: string
    type: object
...
  responses.Meta:
    properties:
      has_next:
        type: boolean
      has_prev:
        type: boolean
      page:
        type: integer
      per_page:
        type: integer
      total:
        type: integer
      total_pages:
        type: integer
    type: object
...
  responses.Response:
    properties:
      data: {}
      error: {}
      message:
        type: string
      meta:
        $ref: '#/definitions/responses.Meta'
      request_id:
        type: string
      success:
        type: boolean
      timestamp:
        type: string
    type: object
...
  responses.TicketResponse:
    properties:
      code:
        type: string
      created_at:
        type: string
      event_id:
        type: string
      holder_email:
        type: string
      holder_name:
        type: string
      holder_phone:
        type: string
      id:
        type: string
      issued_at:
        type: string
      occurrence_id:
        type: string
      order_id:
        type: string
      order_item_id:
        type: string
      scan_location:
        type: string
      status:
        type: string
      ticket_type_id:
        type: string
      transfer_count:
        type: integer
      updated_at:
        type: string
      used_at:
        type: string
      user_id:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@1kulture.com
    name: API Support
    url: https://1kulture.com/support
  description: Enterprise Event Management System API with multi-tenant SaaS architecture
  license:
    name: Proprietary
    url: https://1kulture.com/license
  termsOfService: https://1kulture.com/terms
  title: 1Kulture API
  version: "1.0"
paths:
...
  /brands/dashboard:
    get:
      consumes:
      - application/json
      description: Active partnerships, pending requests, available events.
      produces:
      - application/json
      responses:
        "200":
          description: Dashboard
          schema:
            $ref: '#/definitions/responses.BrandDashboardResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Brand dashboard
      tags:
      - brand-discovery
  /brands/discover:
    get:
      consumes:
      - application/json
      description: Brand-facing discovery with partnership-specific filters.
      parameters:
      - description: Category
        in: query
        name: category
        type: string
      - description: City
        in: query
        name: city
        type: string
      - description: Country
        in: query
        name: country
        type: string
      - description: Partnership type (cash_sponsorship, product_sponsorship, ...)
        in: query
        name: partnership_type
        type: string
      - description: Minimum expected attendees
        in: query
        name: min_audience
        type: integer
      - description: Maximum expected attendees
        in: query
        name: max_audience
        type: integer
      - description: RFC3339 start filter
        in: query
        name: start_from
        type: string
      - description: RFC3339 end filter
        in: query
        name: start_to
        type: string
      - description: Search title/summary
        in: query
        name: search
        type: string
      - description: Page
        in: query
        name: page
        type: integer
      - description: Per page
        in: query
        name: per_page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Events
          schema:
            items:
              $ref: '#/definitions/responses.EventSummaryResponse'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Discover events open for partnerships
      tags:
      - brand-discovery
  /brands/events/{id}:
    get:
      consumes:
      - application/json
      parameters:
      - description: Event ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Event + config
          schema:
            $ref: '#/definitions/responses.Response'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "404":
          description: Not found
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get event detail (with partnership config) for a brand
      tags:
      - brand-discovery
...
  /brands/recommended:
    get:
      consumes:
      - application/json
      description: Uses brand's preferred categories.
      parameters:
      - default: 10
        description: Limit
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Events
          schema:
            items:
              $ref: '#/definitions/responses.EventSummaryResponse'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Recommended events for the brand
      tags:
      - brand-discovery
...
  /events/{id}/checkin:
    get:
      consumes:
      - application/json
      parameters:
      - description: Event ID
        in: path
        name: id
        required: true
        type: string
      - description: Page
        in: query
        name: page
        type: integer
      - description: Per page
        in: query
        name: per_page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Check-ins
          schema:
            items:
              $ref: '#/definitions/responses.CheckInRecordResponse'
            type: array
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List check-ins for an event
      tags:
      - check-in
  /events/{id}/checkin/manual:
    post:
      consumes:
      - application/json
      description: Staff/organizer performs a manual check-in using the ticket code.
      parameters:
      - description: Event ID
        in: path
        name: id
        required: true
        type: string
      - description: Manual
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/requests.ManualCheckInRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Checked in
          schema:
            $ref: '#/definitions/responses.CheckInResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "422":
          description: Validation error
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Manual check-in
      tags:
      - check-in
  /events/{id}/checkin/scan:
    post:
      consumes:
      - application/json
      description: Staff/organizer scans a ticket via its QR payload.
      parameters:
      - description: Event ID
        in: path
        name: id
        required: true
        type: string
      - description: Scan
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/requests.CheckInRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Checked in
          schema:
            $ref: '#/definitions/responses.CheckInResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "422":
          description: Validation error
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Scan ticket QR
      tags:
      - check-in
  /events/{id}/checkin/stats:
    get:
      consumes:
      - application/json
      parameters:
      - description: Event ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Stats
          schema:
            $ref: '#/definitions/responses.CheckInStatsResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Check-in stats
      tags:
      - check-in
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"